Changed Gadget to return 64-bit addresses.

+++ src/Gadget.h
@@ -1,6 +1,9 @@
 #ifndef _GADGET_H_
 #define _GADGET_H_
 
+// roptool
+#include "Types.h"
+
 // std
 #include <string>
 #include <memory>
@@ -12,7 +15,7 @@ class Gadget
         virtual bool parse(const std::string& file) = 0;
         
         virtual const std::string& name(void) = 0;
-        virtual int address(void) = 0;
+        virtual u64 address(void) = 0;
 };
 
 typedef std::shared_ptr<Gadget> GadgetPtr;
